pyproject.toml
[project]
name = "p0"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "matplotlib>=3.10.8",
    "networkx>=3.6.1",
    "numpy>=2.3.5",
    "placo>=0.9.18",
    "rerun-sdk>=0.28.1",
    "yourdfpy>=0.0.58",
]
